+++ apps/views/torrents/upload.php
@@ -69,14 +69,14 @@
             <td class="nowrap"><label>Quality</label></td>
             <td>
                 <div class="row">
-                    <?php foreach (app()->site::getQualityTableList() as $quality => $title): ?>
+                    <?php foreach (app()->site->getQualityTableList() as $quality => $title): ?>
                     <?php if (config('torrent_upload.enable_quality_' . $quality)) : ?>
                     <div class="col-md-3">
                         <div class="input-group">
                             <span class="input-group-addon"><label for="<?= $quality ?>"><?= $title ?></label></span>
                             <select class="form-control" id="<?= $quality ?>" name="<?= $quality ?>">
                                 <option value="0">[Choose One]</option>
-                                <?php foreach (app()->site::ruleQuality($quality) as $q): ?>
+                                <?php foreach (app()->site->ruleQuality($quality) as $q): ?>
                                     <option value="<?= $q['id']; ?>"><?= $q['name']; ?></option>
                                 <?php endforeach; ?>
                             </select>
@@ -96,7 +96,7 @@
                             <span class="input-group-addon"><label for="team">Group</label></span>
                             <select id="team" name="team" class="form-control">
                                 <option value="0" selected>[Choose One]</option>
-                                <?php foreach (app()->site::ruleCanUsedTeam() as $team) : ?>
+                                <?php foreach (app()->site->ruleCanUsedTeam() as $team) : ?>
                                     <option value="<?= $team['id'] ?>"><?= $team['name'] ?></option>
                                 <?php endforeach; ?>
                             </select>
@@ -138,7 +138,7 @@
             <td><input id="tags" name="tags" class="form-control" type="text">
                 <div class="tag-help-block" style="margin-top: 4px">
                     Pinned Tags:
-                    <?php foreach (app()->site::rulePinnedTags() as $tag): ?>
+                    <?php foreach (app()->site->rulePinnedTags() as $tag): ?>
                         <a href="javascript:" class="add-tag label label-outline <?= $tag['class_name'] ?>"><?= $tag['tag'] ?></a>
                     <?php endforeach; ?>
                 </div>